[HC-28] refactor(ingredient):ingredientRepository 파일 위치 변경
